requirements.txt
pyTelegramBotAPI==4.16.1
python-telegram-bot[job-queue]==20.8

python-dotenv==1.0.1
bottle==0.12.19
requests==2.31.0
pytz==2024.1

binance-connector==3.6.0

pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4

SQLAlchemy==2.0.29

google-api-python-client==2.129.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

sentry-sdk==2.3.1

pytest==8.2.2
pytest-asyncio==0.23.7
pytest-mock==3.14.0

redis==5.2.1
websocket-client==1.7.0
websockets==12.0

vaderSentiment==3.3.2
sanpy==0.12.2
joblib==1.5.0
scikit-learn==1.6.1
